requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
alembic==1.13.2
asyncpg==0.29.0
aiosqlite==0.20.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# PDF Parsing
PyMuPDF==1.24.10
pdfplumber==0.11.4

# NLP
spacy==3.7.6
openai==1.47.0

# Payment
iyzipay==1.0.47
stripe==10.11.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
